- Refactoring player lookup by name in the API routes. - Added some helper functions for api routes. - Renamed the /app folder -> /api
