--- app/getRoundsByPlayer.js
@@ -1,8 +1,0 @@
-
-const getRoundsByPlayer = (rounds, name) => {
-  return rounds.filter(round => {
-    return round.PlayerName.toLowerCase() === name.toLowerCase();
-  })
-}
-
-exports.getRoundsByPlayer = getRoundsByPlayer;
