+++ package.json
@@ -10,5 +10,7 @@
     "express": "^4.17.1",
     "nodemon": "^2.0.6",
     "yargs": "^16.1.0"
+  },"scripts": {
+      "start": "nodemon index.js"
   }
 }
